requirements.txt
PyGithub
packaging
wget
celery==5.2.7
gurobipy==10.0.0
numpy==1.24.1
pandas==1.5.3
redis==4.4.2
fastapi==0.89.1
psutil
loguru
